#58 The calculation is now done with a higher precision.
Because of possible rounding errors the calculation of values in is now
done with a higher precision than the result called for. The Evaluator
now has a separate MathContext for the calculation and for the result.

By default, the Paper does set the calculation precision to be twice of
the result precision, but at least 4.